app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch
torchvision
numpy
pillow
opencv-python
pydantic